Add files via upload
This Game is Purely build on pygame just for learning purpose. It's just a prototype to show the power of pygame. In this Game as your score keeps on increasing your speed will also increase with that and you just have to keep yourself safe from that blocks.The Block will be rectangular in shape